apply_change_to_project: Fix removal of simple value that's nil already.
If the applied change removed a value and the existing value is
different from the removed value, it is considered a conflict.
However, this is not true in the case the existing value was already
removed.

In addition, improve messages in case of real conflicts when dealing
with removal of simple attributes.