apply_change_to_project: Fix adding existing subproject.
When a subproject is added, its file reference is used in several
places:
- It's added to a `PBXGroup`.
- It's added to the dictionary of project references under `projectRef`.
- A new `PBXContainerItemProxy` is added and it refers to it under
  `containerPortal`.

Beforehand, when a subproject that already exists was added, `find_file`
would not be able to distinguish between the existing and new reference,
and it might refer to it in all places mentioned above, which would make
an invalid project file.
To distinguish between them, the file reference that is
used is the one that doesn't exist in the project references.