apply_change_to_project: Add condition for treating removal as conflict.
When removing a simple attribute value, Kintsugi will fail if the
existing value doesn't match the value that's to be removed.
This commit introduces another condition: The existing value should
also be different than the added value.
This is to prevent cases where there are two places that describe the
same change, where both of them contain a removal difference. In such
cases the value will be different in the second pass of the same change
because it was already overridden in the first pass.