apply_change_to_project: Fix when old and new values types don't match.
When adding or removing simple values, until now there was an assumption
that the old value and new value have the same type. This is not
necessarily the case, for example a build setting can be an array with
multiple values, but it can also be a string if it has only a single
value.