xcodeproj_extensions: patch to_tree_hash to normalize build settings.
This solves the issue where an array build setting is empty in `base`,
has some string in `theirs` and a different string in `ours` leads to
an exception.
`to_tree_hash` is used for creating the diff between `theirs` and
`base`, and this commit monkey patches it in XCBuildConfiguration to
convert all array build settings to actual arrays (instead of either a
string or array). This makes sure that the diff between `theirs` and
`base` is between arrays, and then `apply_change_to_project` can be
aware that the result should also be an array.