apply_change_to_project: Fix adding target product file multiple times.
If a new target is created, the product that belongs to it was added
twice. The reason for this was that diff contained the product twice:
Once under the main group, and once under the product reference group.
The reason why the tests passed until now, is that array_diff considers
all non-unique elements as a single element, and since the two product
files are identical it returned a empty diff.
The same thing can happen with `PBXGroup` and `PBXVariantGroup`, and in
fact with any object that can be under the main group.

The solution is to not add the same object twice, by ignoring all the
additions that don't appear in the main group or in project references.
References such as `PBXBuildFile` will still be treated as before.