apply change: Fix adding component with non default simple attributes.
When adding a component with simple attributes, these attributes might
have default values that aren't `nil`. Previously, when such a thing
occurred it would result in a `MergeError`, because the removed change
was assumed to be `nil` since the component was just added, but in
practice it would be set to the default value.
The solution is to set the removed change value to the default value.