Kintsugi: Treat files as identical by a subset of their attributes.
Previously, file references were treated as identical only if all the
change that describes them were identical. This was problematic in case
someone moved a file and change one of its attributes that wasn't part
of its identity, like the file encoding. In such a case the file would
not be identified as moved. Now however, it is identified by its path,
name and source tree.