apply_change_to_project: Fix adding file or group to non existent group.
Previously the code would crash because the containing group was `nil`.
Still the code raises a `MergeError` as it isn't clear whether the group
should be created again or the file/group addition should be ignored.
The same is done when trying to move a file or a group to a group that
doesn't exist.