apply_change_to_project: Fix adding file to target with existing name.
The bug was reported here: https://github.com/Lightricks/Kintsugi/issues/104
Basically, when a file with a name that already exists is added to a
target, the search for the file only took into account the filename.
Instead, the file is searched for in `change_source_project` and its
group path is taken from there.

As a sidenote - Adding the UUID would also help, but would create other
sorts of problems because UUIDs are not always consistent between
project saves.