xcodeproj_extensions: fix wrong serialization with swift package target.
We have a build plugin which is a Swift package, it's represented as a
`XCSwiftPackageProductDependency` object. The display name contains a "plugin:"
prefix so when it's serialized it's everywhere. The problem is that Xcode
removes this prefix and uses the package name in the keys of the serialized
objects and in refs to them (in the comments).

Below is an example before and after the change.

before (wrong):
/* Begin XCSwiftPackageProductDependency section */
		19AA76E52A9F557900F1F6A1 /* plugin:SomePackageName */ = {
			isa = XCSwiftPackageProductDependency;
			productName = "plugin:SomePackageName";
		};

after (correct):
/* Begin XCSwiftPackageProductDependency section */
		19AA76E52A9F557900F1F6A1 /* SomePackageName */ = {
			isa = XCSwiftPackageProductDependency;
			productName = "plugin:SomePackageName";
		};