CI: test on multiple ruby versions.
Currently, the tests were running on ruby 2.6 and using latest
RubyGems. This is broken since latest RubyGems support Ruby>3. This is
fine since ruby 2.7 is now deprecated.

Kintsugi runs usually on macOS which comes built-in with ruby 2.6, so
it's worth keep testing on that version. So this commit add tests on
multiple ruby versions.