apply_change_to_project: use attribute_name for building change_path.
When merging the tree, Kintsugi keeps track of where it is, to allow
finding the corresponding component in the source project.
Until this commit, the path was constructed in `add_child_to_component`
by appending the display name of the added component.
When adding a file reference this path was used to traverse the project
tree along each object's attributes.

This caused a bug when the display name and the attribute name
are different, e.g., a target's build configuration list has a display
name of "ConfigurationList" and an attribute name of
"build_configuration_list". The camel-case was converted correctly, but
the merge failed due to not finding the attribute "configuration_list".

The solution to this issue is to use the attribute names to create the
path, except when adding an object to a list, where there is no
attribute and then the display name is used.